+++ src/main/kotlin/com/github/pirocraft/challengingtimer/TimerToolWindowFactory.kt
@@ -1,0 +1,15 @@
+package com.github.pirocraft.challengingtimer
+
+import com.intellij.openapi.project.Project
+import com.intellij.openapi.wm.ToolWindow
+import com.intellij.openapi.wm.ToolWindowFactory
+import com.intellij.ui.content.ContentFactory
+
+class TimerToolWindowFactory : ToolWindowFactory {
+    override fun createToolWindowContent(project: Project, toolWindow: ToolWindow) {
+        val timerToolWindow = TimerToolWindow(toolWindow)
+        val contentFactory = ContentFactory.SERVICE.getInstance()
+        val content = contentFactory.createContent(timerToolWindow.content, "", false)
+        toolWindow.contentManager.addContent(content)
+    }
+}