+++ CHANGELOG.md
@@ -4,4 +4,7 @@
 
 ## [Unreleased]
 ### Added
-- Initial scaffold created from [IntelliJ Platform Plugin Template](https://github.com/JetBrains/intellij-platform-plugin-template)
+- I challenge myself with the default 1:30 timer
+- I pause and resume the timer
+- I restart the timer with a double click
+- I restart the timer with a simple click when it is finished